requirements.txt
Flask==1.0.2
Jinja2==2.10
WTForms==2.1
flask_bootstrap
Werkzeug==0.16.0
flask_uploads
flask_wtf
mysql-connector-python
redis
google-cloud-firestore
google-cloud-storage
